SQL Database Integration

+++ src/data/sql/03_create_train_table.sql
@@ -4,13 +4,13 @@ DROP TABLE IF EXISTS train;
 CREATE TABLE train
 (
     row_id INTEGER,
-    timestamp INTEGER,
+    timestamp NUMERIC(11),
     user_id INTEGER,
     content_id INTEGER,
     content_type_id INTEGER,
     task_container_id INTEGER,
     user_answer INTEGER,
     answered_correctly INTEGER,
-    prior_question_elapsed_time NUMERIC(5),
+    prior_question_elapsed_time NUMERIC(8),
     prior_question_had_explanation CHAR(5)
 );
